Tests/LibWeb/Text/input/document-scrollingElement-quirks-mode.html

<!-- quirks mode -->
<script src="include.js"></script>
<script type="">
    test(() => {        
        println(`document.compatMode: ${document.compatMode}`);
        println(`document.scrollingElement is body element: ${document.scrollingElement === document.body}`);
    });
</script>
