Tests/LibWeb/Text/expected/document-scrollingElement-quirks-mode.txt

document.compatMode: BackCompat
document.scrollingElement is body element: true
